+++ config/app.php
@@ -15,7 +15,7 @@ return [
 
     'name' => env('APP_NAME', 'TaskConnect'),
 
-    'version' => env('APP_VERSION', '1.1.0'),
+    'version' => env('APP_VERSION', '1.2.0'),
 
     /*
     |--------------------------------------------------------------------------